OPTIM: startup: fast unique_id allocation for acl.
pattern_finalize_config() uses an inefficient algorithm which is a
problem with very large configuration files. This affects startup, and
therefore reload time. When haproxy is deployed as a router in a
Kubernetes cluster the generated configuration file may be large and
reloads are frequently occuring, which makes this a significant issue.

The old algorithm is O(n^2)
* allocate missing uids - O(n^2)
* sort linked list - O(n^2)

The new algorithm does not care about making sure all IDs are in
sequence without gaps and is O(n)
* find the highest user allocated uid - O(n)
* allocate automatic uids - O(n)

To prevent overflow, manually assigned UIDs with -u are limited to
INT_MAX>>1, typically 1073741823.

Performance examples, startup time in seconds:

    pat_refs old     new
    1000      0.01  <0.01
    10000     2.9    0.03
    20000    12.1    0.06
    30000    27.3    0.09
    40000    44.5    0.12
    50000    78.6    0.16

Please backport to 1.8 and 2.0.